requirements.txt
pdfplumber
openai
python-dotenv


neo4j
neo4j-graphrag